+++ src/main/java/com/cg/service/ITransferService.java
@@ -2,10 +2,11 @@ package com.cg.service;
 
 import com.cg.model.Transfer;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface ITransferService {
     public List<Transfer> selectAllTransfers();
 
-    public long calculateTotalFees();
+    public BigDecimal calculateTotalFees();
 }
